♻️ : #468 - 마이페이지 로직 분리
